# specs/1-obe-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: OBE System API
  description: |
    Outcome-Based Education Management System API
    
    This API provides endpoints for:
    - Authentication (JWT, Google OAuth, Email OTP)
    - User management with RBAC
    - Course and outcome management
    - Assessment and result tracking
    - Analytics and reporting
  version: 1.0.0
  contact:
    name: OBE System API Support
    email: support@obesystem.com
servers:
  - url: http://localhost:8000/api
    description: Local development server
  - url: https://api.obesystem.com/api
    description: Production server
...
paths:
...
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/google:
    post:
      tags: [Authentication]
      summary: Google OAuth callback
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  description: OAuth authorization code
      responses:
        '200':
          description: OAuth authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/otp/request:
    post:
      tags: [Authentication]
      summary: Request OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/otp/verify:
    post:
      tags: [Authentication]
      summary: Verify OTP and login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
                  pattern: '^\d{6}$'
      responses:
        '200':
          description: OTP verified, login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh JWT token
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Logout (invalidate token)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string

  # User Management Endpoints
...
  /courses/{id}/assessments:
    get:
      tags: [Assessments]
      summary: List assessments for course
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of assessments
          content:
            application/json:
              schema:
                type: object
                properties:
                  assessments:
                    type: array
                    items:
                      $ref: '#/components/schemas/AssessmentResponse'

    post:
      tags: [Assessments]
      summary: Create assessment (Teacher)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssessmentRequest'
      responses:
        '201':
          description: Assessment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /assessments/{id}:
    get:
      tags: [Assessments]
      summary: Get assessment details
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Assessment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Assessments]
      summary: Update assessment (Teacher)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssessmentRequest'
      responses:
        '200':
          description: Assessment updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssessmentResponse'
        '403':
          $ref: '#/components/responses/Forbidden'

    delete:
      tags: [Assessments]
      summary: Delete assessment (Teacher)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Assessment deleted successfully
        '403':
          $ref: '#/components/responses/Forbidden'

  # Result Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password, first_name, last_name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        first_name:
          type: string
        last_name:
          type: string
        role_id:
          type: integer
          description: Default role if not provided
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        user:
          $ref: '#/components/schemas/UserResponse'
    UserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        role:
          $ref: '#/components/schemas/RoleResponse'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
    RoleResponse:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
...
    CLOResponse:
      type: object
      properties:
        id:
          type: integer
        course_id:
          type: integer
        code:
          type: string
        description:
          type: string
        created_at:
          type: string
          format: date-time
...
    CreateAssessmentRequest:
      type: object
      required: [name, type, weight, max_score]
      properties:
        name:
          type: string
        type:
          type: string
          enum: [Quiz, Assignment, Exam, Project]
        description:
          type: string
        weight:
          type: number
          minimum: 0
          maximum: 100
        max_score:
          type: number
        due_date:
          type: string
          format: date-time
        clo_ids:
          type: array
          items:
            type: integer
    UpdateAssessmentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        weight:
          type: number
        max_score:
          type: number
        due_date:
          type: string
          format: date-time
        clo_ids:
          type: array
          items:
            type: integer
    AssessmentResponse:
      type: object
      properties:
        id:
          type: integer
        course_id:
          type: integer
        name:
          type: string
        type:
          type: string
        description:
          type: string
        weight:
          type: number
        max_score:
          type: number
        due_date:
          type: string
          format: date-time
        clos:
          type: array
          items:
            $ref: '#/components/schemas/CLOResponse'
        created_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'